Add function rstring_push
This function pushes another rstring to the end of our rstring.

We reach a point where the API is gonna be something like this:

Functions will have a couple of different suffixes, since C doesn't
support overloading

- (none): involves another rstring
- _str: involves a C-string
- _byte: involves a uint8_t

This is what I can think of now, but this style seems pretty
understandable for me.

Signed-off-by: Ron Shabi <ron@ronsh.net>