Add function rstring_cmp_str
Add the cmp variation for an rstring and a C-string.
This is inlined because this is just an `strcmp' wrapper.

The decision to wrap `strcmp' is because of the optimized nature of
strcmp in many C libraries, and the fact that we can't reject the
strings based on length, since we don't know the C-string's length
beforehand.

Signed-off-by: Ron Shabi <ron@ronsh.net>