Add function rstring_find_first_byte
This is basically a wrapper from memchr with the option to start from an
offset. I changed the rstring_status_t enum to a simple define scheme,
because I don't wanna mess with SIZE_MAX being an enum value when some
compilers might implicitly cast an enum to an int. This is just to be on
the safe side :)

Signed-off-by: Ron Shabi <ron@ronsh.net>